factor out reliable sub logic
